+++ Runtime/Parallax/Parallax.cs
@@ -7,11 +7,20 @@ namespace Pixygon.Micro {
         private ParallaxLayer[] _parallaxLayers;
         
         public void Initialize(Transform player, ParallaxLayerData[] layerData) {
+            if (_parallaxLayers != null)
+                ResetParallax();
             _parallaxLayers = new ParallaxLayer[layerData.Length];
             for (var i = 0; i < layerData.Length; i++) {
                 _parallaxLayers[i] = Instantiate(_parallaxLayerPrefab, transform);
                 _parallaxLayers[i].Initialize(player, layerData[i]);
             }
+        }
+
+        private void ResetParallax() {
+            foreach (var g in _parallaxLayers) {
+                Destroy(g);
+            }
+            _parallaxLayers = null;
         }
         
         private void Update() {